[names-refactor] Deleted One off reducers and actions.
