+++ assets/pages/oneoff-contributions/helpers/ajax.js
@@ -5,8 +5,8 @@
 import { addQueryParamToURL } from 'helpers/url';
 import { routes } from 'helpers/routes';
 
-import { checkoutError } from '../actions/oneoffContributionsActions';
-import type { PageState } from '../reducers/reducers';
+import { checkoutError } from '../oneoffContributionsActions';
+import type { PageState } from '../oneOffContributionsReducers';
 
 
 // ----- Setup ----- //
