+++ assets/pages/oneoff-contributions/oneoffContributions.jsx
@@ -27,11 +27,11 @@ import { init as pageInit } from 'helpers/page/page';
 
 import PaymentMethodsContainer from './components/paymentMethodsContainer';
 import FormFields from './components/formFields';
-import reducer from './reducers/reducers';
-import type { PageState } from './reducers/reducers';
+import reducer from './oneOffContributionsReducers';
+import type { PageState } from './oneOffContributionsReducers';
 import postCheckout from './helpers/ajax';
 
-import { setPayPalButton } from './actions/oneoffContributionsActions';
+import { setPayPalButton } from './oneoffContributionsActions';
 
 // ----- Page Startup ----- //
